Refactoring tolls classes
