+++ src/RehauSku.Assist.csproj
@@ -122,11 +122,12 @@
     <Compile Include="Assistant\RequestModifier.cs" />
     <Compile Include="Assistant\SkuExtensions.cs" />
     <Compile Include="PriceListTools\CombineTool.cs" />
-    <Compile Include="PriceListTools\AbstractPriceListTool.cs" />
+    <Compile Include="PriceListTools\PriceListTool.cs" />
     <Compile Include="PriceListTools\MergeTool.cs" />
     <Compile Include="PriceListTools\PriceList.cs" />
     <Compile Include="PriceListTools\PriceListPosition.cs" />
-    <Compile Include="PriceListTools\SourceFile.cs" />
+    <Compile Include="PriceListTools\Source.cs" />
+    <Compile Include="PriceListTools\Target.cs" />
     <Compile Include="Ribbon\RibbonController.cs" />
     <Compile Include="Assistant\HttpClientUtil.cs" />
     <Compile Include="Assistant\StoreResponse.cs" />
